+++ src/Surfnet/ServiceProviderDashboard/Application/Command/Entity/RequestDeletePublishedEntityCommand.php
@@ -37,17 +37,41 @@ class RequestDeletePublishedEntityCommand implements Command
     private $applicant;
 
     /**
-     * @param string $manageId
+     * @var string
      */
-    public function __construct($manageId, Contact $applicatant)
-    {
-        $this->manageId = $manageId;
-        $this->applicant = $applicatant;
-    }
+    private $issueType;
 
     /**
-     * @return string
+     * @var string
      */
+    private $issueSummaryTranslationKey;
+
+    /**
+     * @var string
+     */
+    private $issueDescriptionTranslationKey;
+
+    /**
+     * @param string $manageId
+     * @param Contact $applicant
+     * @param string $issueType
+     * @param string $issueSummaryTranslationKey
+     * @param string $issueDescriptionTranslationKey
+     */
+    public function __construct(
+        $manageId,
+        Contact $applicant,
+        $issueType,
+        $issueSummaryTranslationKey,
+        $issueDescriptionTranslationKey
+    ) {
+        $this->manageId = $manageId;
+        $this->applicant = $applicant;
+        $this->issueType = $issueType;
+        $this->issueSummaryTranslationKey = $issueSummaryTranslationKey;
+        $this->issueDescriptionTranslationKey = $issueDescriptionTranslationKey;
+    }
+
     public function getManageId()
     {
         return $this->manageId;
@@ -56,5 +80,20 @@ class RequestDeletePublishedEntityCommand implements Command
     public function getApplicant()
     {
         return $this->applicant;
+    }
+
+    public function getIssueType()
+    {
+        return $this->issueType;
+    }
+
+    public function getIssueDescriptionTranslationKey()
+    {
+        return $this->issueDescriptionTranslationKey;
+    }
+
+    public function getIssueSummaryTranslationKey()
+    {
+        return $this->issueSummaryTranslationKey;
     }
 }