+++ src/Surfnet/ServiceProviderDashboard/Infrastructure/Jira/Repository/IssueRepository.php
@@ -77,7 +77,6 @@ class IssueRepository implements TicketServiceInterface
         $manageIdFieldLabel
     ) {
         Assert::stringNotEmpty($projectKey, 'Please set "jira_issue_project_key" in parameters.yml');
-        Assert::stringNotEmpty($issueType, 'Please set "jira_issue_type" in parameters.yml');
         Assert::stringNotEmpty($manageIdFieldName, 'Please set "jira_issue_manageid_fieldname" in parameters.yml');
         Assert::stringNotEmpty(
             $manageIdFieldLabel,