+++ tests/integration/Application/CommandHandler/Entity/PublishEntityProductionCommandHandlerTest.php
@@ -18,25 +18,22 @@
 
 namespace Surfnet\ServiceProviderDashboard\Tests\Integration\Application\CommandHandler\Entity;
 
-use League\Tactician\CommandBus;
+use JiraRestApi\Issue\Issue;
 use Mockery as m;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
 use Mockery\Mock;
 use Psr\Log\LoggerInterface;
-use Surfnet\ServiceProviderDashboard\Application\Command\Entity\PublishEntityCommand;
 use Surfnet\ServiceProviderDashboard\Application\Command\Entity\PublishEntityProductionCommand;
 use Surfnet\ServiceProviderDashboard\Application\CommandHandler\Entity\PublishEntityProductionCommandHandler;
+use Surfnet\ServiceProviderDashboard\Application\Service\TicketService;
 use Surfnet\ServiceProviderDashboard\Domain\Entity\Contact;
 use Surfnet\ServiceProviderDashboard\Domain\Entity\Entity;
-use Surfnet\ServiceProviderDashboard\Domain\Mailer\Message;
 use Surfnet\ServiceProviderDashboard\Domain\Repository\EntityRepository;
 use Surfnet\ServiceProviderDashboard\Domain\Repository\PublishEntityRepository;
-use Surfnet\ServiceProviderDashboard\Infrastructure\DashboardBundle\Factory\MailMessageFactory;
 use Surfnet\ServiceProviderDashboard\Infrastructure\DashboardSamlBundle\Security\Authentication\Token\SamlToken;
 use Surfnet\ServiceProviderDashboard\Infrastructure\DashboardSamlBundle\Security\Identity;
 use Surfnet\ServiceProviderDashboard\Infrastructure\Manage\Exception\PublishMetadataException;
 use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 class PublishEntityProductionCommandHandlerTest extends MockeryTestCase
 {
@@ -57,21 +54,6 @@ class PublishEntityProductionCommandHandlerTest extends MockeryTestCase
     private $publishEntityClient;
 
     /**
-     * @var CommandBus|Mock
-     */
-    private $commandBus;
-
-    /**
-     * @var MailMessageFactory|Mock
-     */
-    private $mailMessageFactory;
-
-    /**
-     * @var TokenStorageInterface|Mock
-     */
-    private $tokenStorage;
-
-    /**
      * @var FlashBagInterface|Mock
      */
     private $flashBag;
@@ -81,27 +63,28 @@ class PublishEntityProductionCommandHandlerTest extends MockeryTestCase
      */
     private $logger;
 
+    /**
+     * @var TicketService
+     */
+    private $ticketService;
+
     public function setUp()
     {
 
         $this->repository = m::mock(EntityRepository::class);
         $this->publishEntityClient = m::mock(PublishEntityRepository::class);
-        $this->commandBus = m::mock(CommandBus::class);
-        $this->mailMessageFactory = m::mock(MailMessageFactory::class);
-        $this->tokenStorage = m::mock(TokenStorageInterface::class);
+        $this->ticketService = m::mock(TicketService::class);
         $this->flashBag = m::mock(FlashBagInterface::class);
         $this->logger = m::mock(LoggerInterface::class);
 
         $this->commandHandler = new PublishEntityProductionCommandHandler(
             $this->repository,
             $this->publishEntityClient,
-            $this->commandBus,
-            $this->mailMessageFactory,
-            $this->tokenStorage,
+            $this->ticketService,
             $this->flashBag,
-            $this->logger
+            $this->logger,
+            'customIssueType'
         );
-
 
         parent::setUp();
     }
@@ -116,12 +99,24 @@ class PublishEntityProductionCommandHandlerTest extends MockeryTestCase
 
         $entity = m::mock(Entity::class);
         $entity
-            ->shouldReceive('getNameNl')
+            ->shouldReceive('getId')
+            ->andReturn('123');
+        $entity
+            ->shouldReceive('getNameEn')
             ->andReturn('Test Entity Name');
-
+        $entity
+            ->shouldReceive('getEntityId')
+            ->andReturn('https://app.example.com/');
         $entity
             ->shouldReceive('setStatus')
             ->with(Entity::STATE_PUBLISHED);
+
+        $issue = m::mock(Issue::class)->makePartial();
+        $issue->key = 'CXT-999';
+
+        $this->ticketService
+            ->shouldReceive('createIssueFrom')
+            ->andReturn($issue);
 
         $this->repository
             ->shouldReceive('findById')
@@ -140,25 +135,13 @@ class PublishEntityProductionCommandHandlerTest extends MockeryTestCase
                 'id' => 123,
             ]);
 
-        $this->commandBus
-            ->shouldReceive('handle')
-            ->once();
-
-        $this->tokenStorage->shouldReceive('getToken')
-            ->andReturn($token);
-
-        $message = m::mock(Message::class);
-
-        $this->mailMessageFactory->shouldReceive('buildPublishToProductionMessage')
-            ->with($entity, $contact)
-            ->andReturn($message);
-
         $this->logger
             ->shouldReceive('info')
-            ->times(3);
+            ->times(4);
 
 
-        $command = new PublishEntityProductionCommand('d6f394b2-08b1-4882-8b32-81688c15c489');
+        $applicant = new Contact('john:doe', 'john@example.com', 'John Doe');
+        $command = new PublishEntityProductionCommand('d6f394b2-08b1-4882-8b32-81688c15c489', $applicant);
         $this->commandHandler->handle($command);
     }
 
@@ -166,7 +149,7 @@ class PublishEntityProductionCommandHandlerTest extends MockeryTestCase
     {
         $entity = m::mock(Entity::class);
         $entity
-            ->shouldReceive('getNameNl')
+            ->shouldReceive('getNameEn')
             ->andReturn('Test Entity Name');
 
         $entity
@@ -200,7 +183,8 @@ class PublishEntityProductionCommandHandlerTest extends MockeryTestCase
             ->shouldReceive('add')
             ->with('error', 'entity.edit.error.publish');
 
-        $command = new PublishEntityProductionCommand('d6f394b2-08b1-4882-8b32-81688c15c489');
+        $applicant = new Contact('john:doe', 'john@example.com', 'John Doe');
+        $command = new PublishEntityProductionCommand('d6f394b2-08b1-4882-8b32-81688c15c489', $applicant);
         $this->commandHandler->handle($command);
     }
 }